* Added support for XML and text files as first class entities in the knowledge base * Added support for JPEG images and generic binary file formats as first class entities in the knowledge base * Added function for reading files into the knowledge base * Added file reading support functions to the Utility namespace * Modified the Multicast and Broadcast transports to support file transfers * Made an appropriate test for file reading/transferring over the network * Updated version to 0.9.16
git-svn-id: http://madara.googlecode.com/svn/trunk@617 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/transport/Message_Header.h
@@ -18,7 +18,7 @@ namespace Madara
     #define MAX_ORIGINATOR_LENGTH       64
     #define KNOWLEDGE_VALUE_TYPE        Madara::Knowledge_Record
 
-    static const int64_t  MAX_PACKET_SIZE = 512000;  // 512kb
+    static const int  MAX_PACKET_SIZE = 512000;  // 512kb
 
     class MADARA_Export Message_Header
     {
